card-finder-mcp/pyproject.toml
[project]
name = "resnet-mcp"
version = "0.1.0"
description = "ResNET Image Comparator MCP Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.12.0",
    "tensorflow>=2.15.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "requests>=2.31.0",
    "opencv-python>=4.8.0",
    "pytest>=7.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["resnet-mcp"]
